backend/mybackend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
clarifai==10.5.4
clarifai-grpc==10.6.7
colorama==0.4.6
contextlib2==21.6.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
googleapis-common-protos==1.63.2
grpcio==1.65.2
idna==3.7
inquirerpy==0.3.4
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==1.26.4
pfzy==0.3.4
pillow==10.4.0
prompt_toolkit==3.0.47
protobuf==5.27.3
Pygments==2.18.0
python-rapidjson==1.19
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
schema==0.7.5
sqlparse==0.5.1
tabulate==0.9.0
tqdm==4.66.4
tritonclient==2.48.0
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
